Convert ID fields from uint16 to int
Changed various ID fields in the API and internal logic from `uint16` to `int` to accommodate larger ID values and simplify type consistency across methods that interface with integer IDs. Updated respective test cases and validated the changes in CRD and controller logic to handle this modification effectively. These changes help ensure better compatibility with systems that might produce IDs beyond the `uint16` limit.

+++ api/v1/customcertificate_types.go
@@ -69,7 +69,7 @@ type CustomCertificateStatus struct {
 	// Important: Run "make" to regenerate code after modifying this file
 
 	// CustomCertificateStatus ID in the Nginx Proxy Manager instance
-	Id *uint16 `json:"id,omitempty"`
+	Id *int `json:"id,omitempty"`
 
 	// Expiration time of the certificate
 	// +optional
